Refactor to Shapes/Shapes2D
