--- Graphics/Shape/Point.cs
@@ -1,10 +1,0 @@
-﻿using System;
-namespace RobustEngine.Graphics.Shape
-{
-    public class Point
-    {
-        public Point()
-        {
-        }
-    }
-}
