+++ Graphics/Shapes/Point2D.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace RobustEngine.Graphics.Shapes
+{
+    public class Point2D
+    {
+        public Point2D()
+        {
+        }
+    }
+}
